refactor transform preflight api
